Module minecraft_json::minecraft::data::conditions

Conditions common to all Criterions.

Structs

Block
Block conditions.

DamageSource
Properties of damage source.

Distance
World distance.

Effect
Effect properties.

Enchantment
Minecraft enchantment.

Entity
Tags common to all entities.

EntityFlags
Predicate Flags to be checked for an entity.

Equipment
Entity equipments.

Fluid
Fluid conditions.

Item
Tags common to all items.

LightningBolt
Lightning bolt properties.

Location
Tags common to all locations.

Player
Player properties.

Statistic
A statistic. Note that unlike when adding scoreboard objectives, the base (for example, minecraft:custom) and the statistic (for example, minecraft:sneak_time) are split and use proper namespaces instead of the dot-notation (minecraft.custom:minecraft.sneak_time).

Enums

AdvancementStatus
Status of an advancement.

GameMode
The game mode of a player.

Type Definitions

PredicatesOrEntity
Either a list of predicates, or an entity.