Module minecraft_json::minecraft::data::predicate

Predicates are technical JSON files that represent the conditions for loot tables, /execute if predicate command, or predicate target selector argument. They are a part of data packs.

Inside a data pack, a predicate is located within data//predicates.

A predicate file may also contain an array of multiple predicate objects, in which case the predicate passes only if all sub-predicates pass.

Enums

Predicate
Predicate.

WhichEntity
Specifies the entity to check for the condition.