Module minecraft_json::minecraft::text

Raw JSON text format.

Structs

ClickEvent
Event upon mouse click on this text component.

Entity
The entity that should be displayed by HoverEvent::ShowEntity.

Formatting
Formatting properties for a TextComponent, can be inherited.

HexColour
Hexadecimal colours.

Interactivity
Interactivity properties for a TextComponent, can be inherited.

Item
The item that should be displayed by HoverEvent::ShowItem.

LegacyEntity
The entity that should be displayed by HoverEvent::ShowEntity.

Score
Displays a score from the scoreboard. Requires component resolution. This component is resolved into a text component containing the scoreboard value.

TextComponentTags
Common properties for TextComponent.

Enums

ClickEventAction
Action performed upon a ClickEvent.

Colour
Text colours, either pre-defined or custom hexadecimal colours.

ColourName
Pre-defined colour names.

HoverEvent
Event upon mouse hover on this text component.

StringLike
A string containing plain text to display directly. Can also be a number or boolean that is displayed directly.

TextComponent
Raw JSON text is made up of text components. There is a single root component, which can have child components, which can have their own children and so on. Components can also have formatting and interactivity added to them, which is inherited by their children.

ValueOrContents
Part of HoverEvent.